requirements.txt
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
docx2python==2.7.3
dominate==2.9.0
et-xmlfile==1.1.0
Flask==3.0.0
Flask-Bootstrap==3.3.7.1
Flask-WTF==1.2.1
google-auth==2.25.2
google-auth-oauthlib==1.2.0
gspread==5.12.3
idna==3.6
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.3
MarkupSafe==2.1.3
oauthlib==3.2.2
openpyxl==3.1.2
Pillow==10.1.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
python-dotenv==1.0.0
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
typing_extensions==4.9.0
urllib3==2.1.0
visitor==0.1.3
Werkzeug==3.0.1
WTForms==3.1.1
